portals/portals-host-exposure.sub.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/stash-utils.sub.js"></script>
<script src="/common/security-features/resources/common.sub.js"></script>
<body>
<script>
  function openPortal(portalSrc) {
    assert_implements("HTMLPortalElement" in self);
    let portal = document.createElement('portal');
    portal.src = portalSrc;
    return portal
  }

  function openPortalAndReceiveMessage(portalSrc) {
    let portal = openPortal(portalSrc)
    let received = new Promise((resolve, reject) => {
      portal.onmessage = resolve;
      document.body.appendChild(portal);
    });
    return received;
  }

  async function openPortalAndReceiveMessageCrossOrigin(portalSrc) {
    let key = guid();
    let portal = openPortal(`${portalSrc}?key=${key}`);
    document.body.appendChild(portal);
    let result = await takeValue(key);
    assert_equals(result, "passed");
  }

  promise_test(() => {
    return openPortalAndReceiveMessage("resources/portal-host.html");
  }, "window.portalHost should be exposed in same-origin portal");

  promise_test(() => {
    return openPortalAndReceiveMessageCrossOrigin(
        "http://{{hosts[alt][www]}}:{{ports[http][0]}}/portals/resources/portal-host.html");
  }, "window.portalHost should be exposed in cross-origin portal");

  promise_test(() => {
    return openPortalAndReceiveMessageCrossOrigin(
        'resources/portal-host-cross-origin-navigate.sub.html');
  }, "window.portalHost should be exposed in portal after cross-origin navigation");

</script>
</body>
